+++ src/client/_assets/js/modules/pledges-and-certificates.js
@@ -303,12 +303,12 @@ function loadCertificate (brand, pledge) {
       executive: {
         name: 'Sherry A. Chris',
         title: 'President and Chief Executive Officer',
-        company: 'Realogy Expansion Brands',
+        company: 'Anywhere Expansion Brands',
         signature: '/img/signatures/sherry-chris.jpg'
       },
       icon: '/img/icons/fhp-icon-bhgre.svg',
       class: 'bhgre',
-      disclaimer: '© 2020 Better Homes and Gardens Real Estate LLC. All rights reserved. Better Homes and Gardens® and the Better Homes and Gardens Real Estate Logo are registered service marks owned by Meredith Corporation and licensed to Better Homes and Gardens Real Estate LLC.Better Homes and Gardens Real Estate LLC fully supports the principles of the Fair Housing Act and the Equal Opportunity Act. Each office is independently owned and operated.'
+      disclaimer: '© 2022 Better Homes and Gardens Real Estate LLC. All rights reserved. Better Homes and Gardens® and the Better Homes and Gardens Real Estate Logo are registered service marks owned by Meredith Corporation and licensed to Better Homes and Gardens Real Estate LLC.Better Homes and Gardens Real Estate LLC fully supports the principles of the Fair Housing Act and the Equal Opportunity Act. Each office is independently owned and operated.'
     },
     'C21': {
       logo: '/img/logos/logo-c21.svg',
@@ -320,7 +320,7 @@ function loadCertificate (brand, pledge) {
       },
       icon: '/img/icons/fhp-icon-c21.svg',
       class: 'c21',
-      disclaimer: '© 2020 Century 21 Real Estate LLC. All rights reserved. CENTURY 21 and the CENTURY 21 Logo are service marks owned by Century 21 Real Estate LLC. Century 21 Real Estate LLC fully supports the principles of the Fair Housing Act and the Equal Opportunity Act. Eachoffice is independently owned and operated.'
+      disclaimer: '© 2022 Century 21 Real Estate LLC. All rights reserved. CENTURY 21 and the CENTURY 21 Logo are service marks owned by Century 21 Real Estate LLC. Century 21 Real Estate LLC fully supports the principles of the Fair Housing Act and the Equal Opportunity Act. Eachoffice is independently owned and operated.'
     },
     'Coldwell Banker': {
       logo: '/img/logos/logo-cb.svg',
@@ -332,7 +332,7 @@ function loadCertificate (brand, pledge) {
       },
       icon: '/img/icons/fhp-icon-cb.svg',
       class: 'cb',
-      disclaimer: '©2020 Coldwell Banker. All Rights Reserved. Coldwell Banker and the Coldwell Banker logos are trademarks of Coldwell Banker Real Estate LLC. The Coldwell Banker® System is comprised of company owned offices which are owned by a subsidiary of Realogy Brokerage Group LLC and franchised offices which are independently owned and operated. The Coldwell Banker System fully supports the principles of the Fair Housing Act and the Equal Opportunity Act.'
+      disclaimer: '©2022 Coldwell Banker. All Rights Reserved. Coldwell Banker and the Coldwell Banker logos are trademarks of Coldwell Banker Real Estate LLC. The Coldwell Banker® System is comprised of company owned offices which are owned by a subsidiary of Anywhere Advisors LLC and franchised offices which are independently owned and operated. The Coldwell Banker System fully supports the principles of the Fair Housing Act and the Equal Opportunity Act.'
     },
     'Corcoran': {
       logo: '/img/logos/logo-corcoran.svg',
@@ -344,31 +344,31 @@ function loadCertificate (brand, pledge) {
       },
       icon: '/img/icons/fhp-icon-corcoran.svg',
       class: 'corcoran',
-      disclaimer: '©2020 Corcoran. All Rights Reserved. Corcoran and the Corcoran logo are trademarks of Corcoran Group LLC. The Corcoran® System is comprised of company-owned offices which are owned by a subsidiary of Realogy Brokerage Group LLC and franchised offices which are independently owned and operated. The Corcoran System fully supports the principles of the Fair Housing Act and the Equal Opportunity Act.'
+      disclaimer: '©2022 Corcoran. All Rights Reserved. Corcoran and the Corcoran logo are trademarks of Corcoran Group LLC. The Corcoran® System is comprised of company-owned offices which are owned by a subsidiary of Anywhere Advisors LLC and franchised offices which are independently owned and operated. The Corcoran System fully supports the principles of the Fair Housing Act and the Equal Opportunity Act.'
     },
     'ERA': {
       logo: '/img/logos/logo-era.svg',
       executive: {
         name: 'Sherry A. Chris',
         title: 'President and Chief Executive Officer',
-        company: 'Realogy Expansion Brands',
+        company: 'Anywhere Expansion Brands',
         signature: '/img/signatures/sherry-chris.jpg'
       },
       icon: '/img/icons/fhp-icon-era.svg',
       class: 'era',
-      disclaimer: '© 2020 ERA Franchise System LLC. All Rights Reserved. ERA® and the ERA Logo are registered service marks owned by ERA Franchise Systems LLC. ERA Franchise Systems LLC fully supports the principles of the Fair Housing Act and the Equal Opportunity Act. Each office is independently owned and operated.'
+      disclaimer: '© 2022 ERA Franchise System LLC. All Rights Reserved. ERA® and the ERA Logo are registered service marks owned by ERA Franchise Systems LLC. ERA Franchise Systems LLC fully supports the principles of the Fair Housing Act and the Equal Opportunity Act. Each office is independently owned and operated.'
     },
-    'Realogy': {
+    'Anywhere': {
       logo: '/img/logos/logo-realogy.svg',
       executive: {
         name: 'Sue Yannaccone',
         title: 'President & CEO',
-        company: 'Realogy Franchise Group',
+        company: 'Anywhere Brands',
         signature: '/img/signatures/sue-yannaccone.png'
       },
       icon: '/img/icons/fhp-icon-realogy.svg',
-      class: 'realogy',
-      disclaimer: '©2020 Realogy Holdings Corp. All Rights Reserved. Realogy Holdings Corp. fully supports the principles of the Fair Housing Act and the Equal Opportunity Act.'
+      class: 'anywhere',
+      disclaimer: '©2022 Anywhere Real Estate Inc. All Rights Reserved. Anywhere Real Estate Inc. fully supports the principles of the Fair Housing Act and the Equal Opportunity Act.'
     },
     'SIR': {
       logo: '/img/logos/logo-sir.svg',
@@ -380,12 +380,12 @@ function loadCertificate (brand, pledge) {
       },
       icon: '/img/icons/fhp-icon-sir.svg',
       class: 'sir',
-      disclaimer: '©2020 Sotheby\'s International Realty. All Rights Reserved. Sotheby\'s International Realty and the Sotheby\'s International Realty logo are trademarks licensed to Sotheby’s International Realty Affiliates LLC and used with permission. The Sotheby\'s International Realty® affiliate network is operated by Sotheby\'s International Realty Affiliates LLC, and the company owned brokerages are operated by Sotheby\'s International Realty, Inc. Both entities are subsidiaries of Realogy Holdings Corp. Both Sotheby\'s International Realty Affiliates LLC and Sotheby\'s International Realty, Inc. fully support the principles of the Fair Housing Act and the Equal Opportunity Act.'
+      disclaimer: '©2022 Sotheby\'s International Realty. All Rights Reserved. Sotheby\'s International Realty and the Sotheby\'s International Realty logo are trademarks licensed to Sotheby’s International Realty Affiliates LLC and used with permission. The Sotheby\'s International Realty® affiliate network is operated by Sotheby\'s International Realty Affiliates LLC, and the company owned brokerages are operated by Sotheby\'s International Realty, Inc. Both entities are subsidiaries of Anywhere Real Estate Inc. Both Sotheby\'s International Realty Affiliates LLC and Sotheby\'s International Realty, Inc. fully support the principles of the Fair Housing Act and the Equal Opportunity Act.'
     }
   }
 
-  // In case of unrecognized brand, use Realogy.
-  const certificateBrand = brands.hasOwnProperty(brand) ? brand : 'Realogy'
+  // In case of unrecognized brand, use Anywhere.
+  const certificateBrand = brands.hasOwnProperty(brand) ? brand : 'Anywhere'
 
   // Add brand certificate class
   body.classList.add(`certificate_${brands[certificateBrand].class}`)